Table title and subtitle dynamically populating. transforming index page into home component for reuse on lists pages
